+++ R/data.R
@@ -1,7 +1,7 @@
-#' Eon data from the International Commission on Stratigraphy (v2021/05)
+#' Eon data from the International Commission on Stratigraphy (v2022/02)
 #'
 #' A dataset containing the boundary ages, abbreviations, and colors for the eons of the Geologic Time Scale.
-#' Based on The ICS International Chronostratigraphic Chart (v2021/05), by Cohen, Finney, Gibbard, and Fan.
+#' Based on The ICS International Chronostratigraphic Chart (v2022/02), by Cohen, Finney, Gibbard, and Fan.
 #'
 #' @format A data frame with 3 rows and 5 variables:
 #' \describe{
@@ -15,10 +15,10 @@
 #' @source \url{https://stratigraphy.org} via \url{https://macrostrat.org/api/v2/defs/intervals?timescale=international\%20eons}
 "eons"
 
-#' Era data from the International Commission on Stratigraphy (v2021/05)
+#' Era data from the International Commission on Stratigraphy (v2022/02)
 #'
 #' A dataset containing the boundary ages, abbreviations, and colors for the eras of the Geologic Time Scale.
-#' Based on The ICS International Chronostratigraphic Chart (v2021/05), by Cohen, Finney, Gibbard, and Fan.
+#' Based on The ICS International Chronostratigraphic Chart (v2022/02), by Cohen, Finney, Gibbard, and Fan.
 #'
 #' @format A data frame with 10 rows and 5 variables:
 #' \describe{
@@ -32,10 +32,10 @@
 #' @source \url{https://stratigraphy.org} via \url{https://macrostrat.org/api/v2/defs/intervals?timescale=international\%20eras}
 "eras"
 
-#' Period data from the International Commission on Stratigraphy (v2021/05)
+#' Period data from the International Commission on Stratigraphy (v2022/02)
 #'
 #' A dataset containing the boundary ages, abbreviations, and colors for the periods of the Geologic Time Scale.
-#' Based on The ICS International Chronostratigraphic Chart (v2021/05), by Cohen, Finney, Gibbard, and Fan.
+#' Based on The ICS International Chronostratigraphic Chart (v2022/02), by Cohen, Finney, Gibbard, and Fan.
 #'
 #' @format A data frame with 22 rows and 5 variables:
 #' \describe{
@@ -49,10 +49,10 @@
 #' @source \url{https://stratigraphy.org} via \url{https://macrostrat.org/api/v2/defs/intervals?timescale=international\%20periods}
 "periods"
 
-#' Epoch data from the International Commission on Stratigraphy (v2021/05)
+#' Epoch data from the International Commission on Stratigraphy (v2022/02)
 #'
 #' A dataset containing the boundary ages, abbreviations, and colors for the epochs of the Geologic Time Scale.
-#' Based on The ICS International Chronostratigraphic Chart (v2021/05), by Cohen, Finney, Gibbard, and Fan.
+#' Based on The ICS International Chronostratigraphic Chart (v2022/02), by Cohen, Finney, Gibbard, and Fan.
 #'
 #' @format A data frame with 34 rows and 5 variables:
 #' \describe{
@@ -66,10 +66,10 @@
 #' @source \url{https://stratigraphy.org} via \url{https://macrostrat.org/api/v2/defs/intervals?timescale=international\%20epochs}
 "epochs"
 
-#' Stage data from the International Commission on Stratigraphy (v2021/05)
+#' Stage data from the International Commission on Stratigraphy (v2022/02)
 #'
 #' A dataset containing the boundary ages, abbreviations, and colors for the stages of the Geologic Time Scale.
-#' Based on The ICS International Chronostratigraphic Chart (v2021/05), by Cohen, Finney, Gibbard, and Fan.
+#' Based on The ICS International Chronostratigraphic Chart (v2022/02), by Cohen, Finney, Gibbard, and Fan.
 #'
 #' @format A data frame with 102 rows and 5 variables:
 #' \describe{