+++ NEWS.md
@@ -1,3 +1,6 @@
+# development version
+
+
 # deeptime 0.2.2
 
 * Skip tests when suggested packages are unavailable