+++ NEWS.md
@@ -1,3 +1,5 @@
+# development version
+
 # deeptime 0.2.3
 
 * Updated built-in timescales to 2022/02 version of GTS